--- www/src/lib/rootIntersectionObserver/index.ts
@@ -1,1 +1,0 @@
-export * from "./rootIntersectionObserver";
