+++ www/src/lib/sharedIntersectionObserver/sharedIntersectionObserver.ts
